requirements.txt
Flask==3.1.1
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
boto3==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
